+++ BulbChipSTM32C071FBPx/Core/Inc/microlight/json/command_parser.h
@@ -47,6 +47,8 @@
  * }
  */
 
-void parseJson(const char buf[], uint32_t count, CliInput *input);
+#include <stddef.h>
+
+void parseJson(const char buf[], size_t count, CliInput *input);
 
 #endif /* INC_COMMAND_PARSER_H_ */
