+++ BulbChipSTM32C071FBPx/Core/Inc/usb_dependencies.h
@@ -9,11 +9,12 @@
 #define INC_USB_DEPENDENCIES_H_
 
 #include <stdbool.h>
+#include <stddef.h>
 #include <stdint.h>
 
-void usbWriteToSerial(const char *buf, int count);
+void usbWriteToSerial(const char *buf, size_t count);
 
 // returns bytes read if a buffer has read in a full line terminated by \n, 0 otherwise.
-int usbCdcReadTask(char usbBuffer[], int bufferLength);
+int32_t usbCdcReadTask(char usbBuffer[], size_t bufferLength);
 
 #endif /* INC_USB_DEPENDENCIES_H_ */
